src/requirements.txt
decorator==4.0.10
docutils==0.12
fake-factory==0.5.7
ipaddress==1.0.16
lxml==3.6.0
Pygments==2.1.3
pymssql==2.1.2
python-dateutil==2.5.3
pythonnet==2.0.0
PyYAML==3.13
robotframework==3.0
robotframework-databaselibrary==0.6
robotframework-faker==3.0.0
robotframework-medley==0.3
robotframework-ride==1.5.2.1
robotframework-selenium2library==1.7.4
selenium==2.53.4
six==1.10.0
wrapt==1.10.8
